specconv: temporarily allow userns path and mapping if they match
(This is a cherry-pick of b9c58080a3f6e2104668849ec9cb773f588f2f4c.)

It turns out that the error added in commit 1a1476792262 ("configs:
disallow ambiguous userns and timens configurations") causes issues with
containerd and CRIO because they pass both userns mappings and a userns
path.

These configurations are broken, but to avoid the regression in this one
case, output a warning to tell the user that the configuration is
incorrect but we will continue to use it if and only if the configured
mappings are identical to the mappings of the provided namespace.

Fixes: 1a1476792262 ("configs: disallow ambiguous userns and timens configurations")
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>