Revert "increase memory.max in cgroups.bats"
This reverts commit 65a1074c75123afbe1d59e6ca2abc683e9d0d9dc.

we need (#4020) because of (#3931), at that time, we removed the bindfd logic, and the
memfd logic will use more memory than before, but we have not yet moved binary clone from
runc init to runc parent process, so we need to increase memory limit in CI.
As we have moved the runc binary clone logic from runc init to runc parent process in
(#3987), so the memory usage of binary clone will not be included in container's memory
cgroup accounting. Now we can support run a simple container with lower memory usage the
same as before.

Signed-off-by: lfbzhm <lifubang@acmcoder.com>